pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=2.0.0",
    "twilio>=9.5.1",
]
